requirements.txt
Orange3>=3.3.5
hyper
pillow
requests
